server_end/requirements.txt
click==6.7
dlib==19.10.0
face-recognition==1.2.2
face-recognition-models==0.3.0
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
numpy==1.14.3
Pillow==5.1.0
Werkzeug==0.14.1
asn1crypto==0.24.0
cffi==1.11.5
click==6.7
cryptography==2.2.2
dlib==19.10.0
face-recognition==1.2.2
face-recognition-models==0.3.0
Flask==1.0.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
numpy==1.14.3
Pillow==5.1.0
pycparser==2.18
pyOpenSSL==17.5.0
six==1.11.0
Werkzeug==0.14.1
